Rename `nwbib` controllers and routes packages to `rpb` (RPB-272)
