+++ app/controllers/rpb/WikidataLocations.java
@@ -1,4 +1,4 @@
-package controllers.nwbib;
+package controllers.rpb;
 
 import java.io.File;
 import java.io.FileInputStream;
