+++ app/views/browse_classification.scala.html
@@ -13,13 +13,13 @@
     @tags.browse_list(tops, subs, t, placeholder)
     @if(t=="Wikidata"){
         <p style="margin-top:10px">@(subs.values.flatMap((vals: List[JsonNode]) => vals).size + Json.parse(tops).asOpt[Seq[JsValue]].getOrElse(Seq()).size - 1) Orte | 
-            <a href='@nwbib.routes.Application.classification("WikidataImport")'>aus Wikidata aktualisieren
+            <a href='@rpb.routes.Application.classification("WikidataImport")'>aus Wikidata aktualisieren
             <span class="glyphicon glyphicon-repeat"></span></a>
         </p>
     } else {
         <p>
-          <a href="@if(t=="Raumsystematik"){@nwbib.routes.Application.spatial("register")}else{@nwbib.routes.Application.subjects("register")}"><span class="glyphicon glyphicon-sort-by-alphabet"></span> A-Z Register</a> | 
-          <a href="@if(t=="Raumsystematik"){@nwbib.routes.Application.spatialTtl()}else{@nwbib.routes.Application.subjectsTtl()}"><span class="glyphicon glyphicon-download-alt"></span> SKOS-Datei</a>
+          <a href="@if(t=="Raumsystematik"){@rpb.routes.Application.spatial("register")}else{@rpb.routes.Application.subjects("register")}"><span class="glyphicon glyphicon-sort-by-alphabet"></span> A-Z Register</a> | 
+          <a href="@if(t=="Raumsystematik"){@rpb.routes.Application.spatialTtl()}else{@rpb.routes.Application.subjectsTtl()}"><span class="glyphicon glyphicon-download-alt"></span> SKOS-Datei</a>
         </p>
     }
 }
