+++ app/views/browse_register.scala.html
@@ -6,5 +6,5 @@
 
 @main("RPB - " + placeholder.replace("filtern","").trim) {
     @tags.browse_list(classification, null, t, placeholder)
-    <p><a href="@if(t=="Raumsystematik"){@nwbib.routes.Application.spatial()}else{@nwbib.routes.Application.subjects()}"><span class="octicon octicon-list-ordered"></span> Systematik</a></p>
+    <p><a href="@if(t=="Raumsystematik"){@rpb.routes.Application.spatial()}else{@rpb.routes.Application.subjects()}"><span class="octicon octicon-list-ordered"></span> Systematik</a></p>
 }
