+++ app/views/classification.scala.html
@@ -4,9 +4,9 @@
 
 @main("NWBib - Systematiken") {
 <dl>
-    <dt><a href="@nwbib.routes.Application.classification("Raumsystematik")">Raumsystematik</a></dt>
+    <dt><a href="@rpb.routes.Application.classification("Raumsystematik")">Raumsystematik</a></dt>
     <dd>Raumsystematik der Nordrhein-Westfälischen Bibliographie</dd>
-    <dt><a href="@nwbib.routes.Application.classification("Sachsystematik")">Sachsystematik</a></dt>
+    <dt><a href="@rpb.routes.Application.classification("Sachsystematik")">Sachsystematik</a></dt>
     <dd>Sachsystematik der Nordrhein-Westfälischen Bibliographie</dd>
 </dl>
 }
