+++ app/views/details.scala.html
@@ -5,8 +5,8 @@
 @import helper._
 @import tags._
 @import views.TableRow
-@import controllers.nwbib.Lobid
-@import controllers.nwbib.Application
+@import controllers.rpb.Lobid
+@import controllers.rpb.Application
 @import play.api.libs.json.Json
 @import play.api.libs.json.JsValue
 @import play.api.libs.json.JsArray
@@ -20,7 +20,7 @@
 @pageLink(last: Seq[String], key: String, label: String, tip: String) = {
   @defining((last.indexOf(session(key))-1,last.indexOf(session(key))+1)) { case (prev,next) =>
   <li class="@if(session(key).isEmpty){disabled}">
-    <a title="@tip" href=@if(session(key).isEmpty) {"#"} else {"@nwbib.routes.Application.show(session(key),"")"}>
+    <a title="@tip" href=@if(session(key).isEmpty) {"#"} else {"@rpb.routes.Application.show(session(key),"")"}>
       @session(session(key), (if(prev >=0) last(prev) else "") + "," + (if(next < last.size()) last(next) else ""))
       @Html(label)
     </a>
@@ -53,13 +53,13 @@
                 @pageLink(last, "prev", "&larr;", "Voriger Treffer")
                 @defining(session("lastSearchUrl")){ lastSearchUrl =>
                   <li><a href="@lastSearchUrl">Treffer @(last.indexOf((doc\\"rpbId")(0).as[String])+1) von @last.size in der 
-                  @if(lastSearchUrl==nwbib.routes.Application.showStars().toString()) {Merkliste} else {letzten Trefferliste}</a></li>
+                  @if(lastSearchUrl==rpb.routes.Application.showStars().toString()) {Merkliste} else {letzten Trefferliste}</a></li>
                 }
                 @pageLink(last, "next", "&rarr;", "Nächster Treffer")
               </ul>
             </nav>
             <script>
-              if(document.referrer.indexOf('@nwbib.routes.Application.index()') < 0){
+              if(document.referrer.indexOf('@rpb.routes.Application.index()') < 0){
                 $("#pager").hide();
               }
             </script>
