+++ app/views/index.scala.html
@@ -5,13 +5,13 @@
 @import helper._
 @import play.api.libs.json.Json
 @import play.api.libs.json.JsValue
-@import controllers.nwbib.Lobid
+@import controllers.rpb.Lobid
 
 @main("RPB") {
     @tags.search_form("")
     <div class="row">
 	    @*<div class="col-md-7 intro">
-	        @defining(controllers.nwbib.Lobid.getTotalHits("").get(Lobid.API_TIMEOUT)){ hits =>
+	        @defining(controllers.rpb.Lobid.getTotalHits("").get(Lobid.API_TIMEOUT)){ hits =>
 	       <p class="lead">
            Die Nordrhein-Westfälische Bibliographie (NWBib) verzeichnet die Literatur über das Land Nordrhein-Westfalen, seine Regionen, 
            Orte und Persönlichkeiten, Literatur aus allen Lebens- und Wissensbereichen in Geschichte und Gegenwart.
@@ -20,7 +20,7 @@
            Die NWBib ist eine der umfangreichsten Regionalbibliographien Deutschlands. Sie erschließt nicht nur Bücher und Zeitschriften, 
            sondern auch Aufsätze und andere Medien wie etwa Karten, DVDs, Hörbücher und elektronische Publikationen. Zum raschen Einstieg 
            nutzen Sie die Eingabezeile oder die Karte mit 
-           @if(map == "kreise") {Kreisen (oder <a href='@nwbib.routes.Application.index(map="gemeinden")'>Gemeinden</a>)} else {Gemeinden (oder <a href='@nwbib.routes.Application.index(map="kreise")'>Kreisen</a>)}.
+           @if(map == "kreise") {Kreisen (oder <a href='@rpb.routes.Application.index(map="gemeinden")'>Gemeinden</a>)} else {Gemeinden (oder <a href='@rpb.routes.Application.index(map="kreise")'>Kreisen</a>)}.
 	       </p>
 	       <p>
 	       Für die präzise Recherche steht Ihnen die
