+++ app/views/register.scala.html
@@ -4,9 +4,9 @@
 
 @main("NWBib - Register") {
 <dl>
-    <dt><a href="@nwbib.routes.Application.register("Raumsystematik")">Register zur Raumsystematik</a></dt>
+    <dt><a href="@rpb.routes.Application.register("Raumsystematik")">Register zur Raumsystematik</a></dt>
     <dd>Alphabetisches Register zur Raumsystematik der Nordrhein-Westfälischen Bibliographie</dd>
-    <dt><a href="@nwbib.routes.Application.register("Sachsystematik")">Register zur Sachsystematik</a></dt>
+    <dt><a href="@rpb.routes.Application.register("Sachsystematik")">Register zur Sachsystematik</a></dt>
     <dd>Alphabetisches Register zur Sachsystematik der Nordrhein-Westfälischen Bibliographie</dd>
 </dl>
 }
