+++ app/views/search.scala.html
@@ -4,10 +4,10 @@
 
 @import helper._
 @import tags._
-@import controllers.nwbib.Lobid
-@import controllers.nwbib.Application
-@import controllers.nwbib.Classification
-@import controllers.nwbib.WikidataLocations
+@import controllers.rpb.Lobid
+@import controllers.rpb.Application
+@import controllers.rpb.Classification
+@import controllers.rpb.WikidataLocations
 @import play.api.libs.json.Json
 @import play.api.libs.json.JsValue
 @import play.api.libs.json.JsArray
@@ -16,17 +16,17 @@
   <nav>
     <ul class="pagination">
       <li class="previous @if(from==0){disabled}">
-        <a href="@if(from==0){#} else {@nwbib.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,from-size,size,owner,t,sortParam,location=location,word=word,corporation=corporation,raw=raw)}">&larr;</a>
+        <a href="@if(from==0){#} else {@rpb.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,from-size,size,owner,t,sortParam,location=location,word=word,corporation=corporation,raw=raw)}">&larr;</a>
       </li>
       @defining((((from+1)/size)+1,(if(allHits%size==0) allHits/size else allHits/size+1).toInt)) { case (currentPage,lastPage) =>
           @defining(Math.min(Math.max(1,currentPage-4)+9,lastPage)) { toPage =>
               @for(i <- Math.max(1,toPage-9) to toPage){
-                  <li @if(currentPage==i){class="active"}><a href="@nwbib.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,(i*size)-size,size,owner,t,sortParam,location=location,word=word,corporation=corporation,raw=raw)">@i</a></li>
+                  <li @if(currentPage==i){class="active"}><a href="@rpb.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,(i*size)-size,size,owner,t,sortParam,location=location,word=word,corporation=corporation,raw=raw)">@i</a></li>
               }
           }
       }
       <li class="next @if(from+size >= allHits){disabled}">
-        <a href="@if(from+size >= allHits){#} else {@nwbib.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,from+size,size,owner,t,sortParam,location=location,word=word,corporation=corporation,raw=raw)}">&rarr;</a>
+        <a href="@if(from+size >= allHits){#} else {@rpb.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,from+size,size,owner,t,sortParam,location=location,word=word,corporation=corporation,raw=raw)}">&rarr;</a>
       </li>
     </ul>
   </nav>
@@ -34,13 +34,13 @@
 
 @sort(param: String, label: String)={
 <li role="tab" @if(sortParam==param){class="active" aria-selected="true"} else {aria-selected="false" tabindex="-1" aria-controls="@label"}>
-  <a href="@nwbib.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,from,size,owner,t,param,location=location,word=word,corporation=corporation,raw=raw)">@label</a>
+  <a href="@rpb.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,from,size,owner,t,param,location=location,word=word,corporation=corporation,raw=raw)">@label</a>
 </li>
 }
 
 @pageLink(num: Int)={
 <li role="tab" @if(size==num){class="active" aria-selected="true"} else {aria-selected="false" tabindex="-1" aria-controls="@num"}>
-  <a href="@nwbib.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,from,num,owner,t,sortParam,location=location,word=word,corporation=corporation,raw=raw)">@num</a>
+  <a href="@rpb.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,from,num,owner,t,sortParam,location=location,word=word,corporation=corporation,raw=raw)">@num</a>
 </li>
 }
 
@@ -93,7 +93,7 @@
                  <div class="panel-body" style="text-align:center">
                  <div class="row">
                     <div class="col-md-1" style="text-align: left; padding-left: 4px;">
-                    <a rel='alternate' type='application/ld+json' href='@nwbib.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,from,size,owner,t,sortParam,location=location,word=word,corporation=corporation,raw=raw,format="json")'>
+                    <a rel='alternate' type='application/ld+json' href='@rpb.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,from,size,owner,t,sortParam,location=location,word=word,corporation=corporation,raw=raw,format="json")'>
                         <img class='json-ld-icon' src='@routes.Assets.at("images/json-ld.png")' title='JSON-LD für diese Suchanfrage'>
                     </a>
                     </div>
@@ -106,9 +106,9 @@
                     <div class="col-md-1" style="text-align: right; padding-right: 4px;">
                       @defining((Json.parse(result)\\"rpbId").map(_.as[String]).mkString(",")) { ids =>
                       @if(ids.split(",").forall(Application.currentlyStarred().split(" ").contains)) {
-                        <a href='@nwbib.routes.Application.clearStars(ids)' title="Alle Titel dieser Seite von Merkliste entfernen"><span class="glyphicon glyphicon-star"></span></a>
+                        <a href='@rpb.routes.Application.clearStars(ids)' title="Alle Titel dieser Seite von Merkliste entfernen"><span class="glyphicon glyphicon-star"></span></a>
                       } else {
-                        <a href='@nwbib.routes.Application.starAll(ids)' title="Alle Titel dieser Seite zur Merkliste hinzufügen"><span class="glyphicon glyphicon-star-empty"></span></a> 
+                        <a href='@rpb.routes.Application.starAll(ids)' title="Alle Titel dieser Seite zur Merkliste hinzufügen"><span class="glyphicon glyphicon-star-empty"></span></a> 
                       }
                       }
                     </div>
