+++ app/views/tags/breadcrumb.scala.html
@@ -1,9 +1,9 @@
 @* Copyright 2019 Fabian Steeg, hbz. Licensed under the GPLv2 *@
 
-@import controllers.nwbib.Lobid
-@import controllers.nwbib.Classification
-@import controllers.nwbib.Classification.Type
-@import controllers.nwbib.WikidataLocations
+@import controllers.rpb.Lobid
+@import controllers.rpb.Classification
+@import controllers.rpb.Classification.Type
+@import controllers.rpb.WikidataLocations
 
 @(rpbsubject: String, rpbspatial: String, person:String="", name:String="", subject:String="", id:String="", publisher:String="", issued:String="", medium: String="", owner: String="", t: String="", location: String="", word: String="", corporation: String="", raw: String="", search: Boolean = false)
 
@@ -11,22 +11,22 @@
 
 @labels(uri: String) = { @defining(Lobid.facetLabel(Seq(uri),"","")) { label =>
 	@if(Lobid.isGnd(uri)) {
-		<a title='@toInternalGndId(uri.split("#").last) @label' href='@if(label.startsWith("http")){@label}else{@nwbib.routes.Application.spatial()#@toInternalGndId(uri.split("#").last)}'>
+		<a title='@toInternalGndId(uri.split("#").last) @label' href='@if(label.startsWith("http")){@label}else{@rpb.routes.Application.spatial()#@toInternalGndId(uri.split("#").last)}'>
 			<span style='color: #777777'>@Classification.notation(uri, Type.SPATIAL)</span> @label
 		</a>
 	} else {
-		<a title='@Classification.shortId(uri) @Lobid.facetLabel(Seq(uri),"","")' href='@if(uri.contains("spatial")){@nwbib.routes.Application.spatial()}else{@nwbib.routes.Application.subjects()}#@(if(uri.contains("#")){"n"+uri.split("#")(1).substring(1)}else{uri})'>
+		<a title='@Classification.shortId(uri) @Lobid.facetLabel(Seq(uri),"","")' href='@if(uri.contains("spatial")){@rpb.routes.Application.spatial()}else{@rpb.routes.Application.subjects()}#@(if(uri.contains("#")){"n"+uri.split("#")(1).substring(1)}else{uri})'>
 			@*<span style='color: #777777'>@Classification.shortId(uri)</span> *@@if(!uri.endsWith("#n9")){@label}
 		</a>
 	}
 }}
 
 @if(!rpbsubject.isEmpty && !rpbsubject.contains(",") && Seq(rpbspatial, person, name, subject, id, publisher, issued, medium, owner, t, location, word, corporation, raw).forall(_.isEmpty)){
-	@if(search){in <a href="@nwbib.routes.Application.subjects()"> Sachgebiete</a> &gt;}
+	@if(search){in <a href="@rpb.routes.Application.subjects()"> Sachgebiete</a> &gt;}
 	@labels(rpbsubject)
 } else {
 	@if(!rpbspatial.isEmpty && !rpbspatial.contains(",") && Seq(rpbsubject, person, name, subject, id, publisher, issued, medium, owner, t, location, word, corporation, raw).forall(_.isEmpty)){
-		@if(search){in <a href="@nwbib.routes.Application.spatial()"> Regionen</a> &gt;}
+		@if(search){in <a href="@rpb.routes.Application.spatial()"> Regionen</a> &gt;}
 		@if(!rpbspatial.isEmpty && !rpbspatial.startsWith("https://rpb.lobid.org/spatial#n")) {
 			@labels(rpbspatial)
 		} else {
