+++ app/views/tags/facets.scala.html
@@ -1,6 +1,6 @@
 @(q:String, person:String, name:String, subject:String, id:String, publisher:String, issued:String, medium: String, rpbspatial: String, rpbsubject: String, owner: String, t: String, sortParam: String, location: String, word: String, corporation: String, raw: String, from: Int, size: Int)
 
-@import controllers.nwbib.Application
+@import controllers.rpb.Application
 @import helper._
 
 <script>
@@ -107,7 +107,7 @@
 @facetLabel(current: String, label: String, activeCountId: String = "", currentCount: String = "") = {
   @if(!current.isEmpty){
      <span class="badge" id="facet-filter"><span id="@activeCountId">@currentCount</span> @label</span>
-     <a href="@nwbib.routes.Application.search(q,person,name,if (urisOnly(subject)==current) "" else subject,id,publisher,if(issued==current) "" else issued,if (medium==current) "" else medium,if(rpbspatial==current) "" else rpbspatial,if(rpbsubject==current) "" else rpbsubject,from,size,if (owner==current) "" else owner,if (t==current) "" else t,sortParam,location=(if (location==current) "" else location),word=word,corporation=corporation,raw=(if (raw.split(":").last==current) "" else raw)).toString" title="Filter entfernen"><span class="octicon octicon-x"></span></a> 
+     <a href="@rpb.routes.Application.search(q,person,name,if (urisOnly(subject)==current) "" else subject,id,publisher,if(issued==current) "" else issued,if (medium==current) "" else medium,if(rpbspatial==current) "" else rpbspatial,if(rpbsubject==current) "" else rpbsubject,from,size,if (owner==current) "" else owner,if (t==current) "" else t,sortParam,location=(if (location==current) "" else location),word=word,corporation=corporation,raw=(if (raw.split(":").last==current) "" else raw)).toString" title="Filter entfernen"><span class="octicon octicon-x"></span></a> 
    }
 }
 
@@ -129,7 +129,7 @@
 <div class="row facet"><div class="col-md-12"><canvas id="issued-chart"></canvas></div></div>
   <div aria-hidden="true" id="slider-range"></div>
   <input type="text" id="issued-range" readonly value="Lade Erscheinungsjahre..." aria-label="Erscheinungsjahre">
-  <div id="issued-div"> <a id="issued-link" style="display:none" href="@nwbib.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,from,size,owner,t,sortParam,location=location,word=word,corporation=corporation,raw=raw)">Anwenden</a> </div>
+  <div id="issued-div"> <a id="issued-link" style="display:none" href="@rpb.routes.Application.search(q,person,name,subject,id,publisher,issued,medium,rpbspatial,rpbsubject,from,size,owner,t,sortParam,location=location,word=word,corporation=corporation,raw=raw)">Anwenden</a> </div>
   </div>
   <br/>
   <div class="row facet"><div class="col-md-12">@facets("Regionen", Application.NWBIB_SPATIAL_FIELD, rpbspatial)</div></div>
