+++ app/views/tags/issued_facet.scala.html
@@ -1,6 +1,6 @@
 @* Copyright 2015 Fabian Steeg, hbz. Licensed under the GPLv2 *@
 
-@import controllers.nwbib.Application
+@import controllers.rpb.Application
 
 @(issued: String)
