+++ app/views/tags/items_map.scala.html
@@ -10,7 +10,7 @@
 @import ExecutionContext.Implicits.global
 @import scala.concurrent.duration._
 @import play.api.Play.current
-@import controllers.nwbib._
+@import controllers.rpb._
 @import play.api.cache.Cache
 
 @string(value: JsValue) = { @value.asOpt[String].getOrElse("--") }
