+++ app/views/tags/result_field.scala.html
@@ -4,7 +4,7 @@
 
 @import play.api.libs.json._
 @import java.util.Optional
-@import controllers.nwbib.Lobid
+@import controllers.rpb.Lobid
 
 @jsonVal(p: String) = @{
 	node match {
