+++ app/views/tags/result_short.scala.html
@@ -3,10 +3,10 @@
 @(id:String, doc: play.api.libs.json.JsValue, i: Int = -1, hits: Seq[String] = Seq())
 
 @import play.api.libs.json._
-@import controllers.nwbib.Lobid
+@import controllers.rpb.Lobid
 @import play.mvc.Controller.session
 @import play.cache.Cache
-@import controllers.nwbib.Application
+@import controllers.rpb.Application
 
 @optional(doc: JsValue, key: String, sep: String, alt: String)=@{
     val other = doc \\ key
@@ -23,10 +23,10 @@
   <td>
     @icon_span((doc \ "type").asOpt[Seq[String]].getOrElse(Seq()), Application.TYPE_FIELD)<br/>
     @if((doc \ "medium" \\ "id").nonEmpty) {
-      @icon_span(Seq((doc \ "medium" \\ "id").last.asOpt[String].getOrElse("")), controllers.nwbib.Application.MEDIUM_FIELD)
+      @icon_span(Seq((doc \ "medium" \\ "id").last.asOpt[String].getOrElse("")), controllers.rpb.Application.MEDIUM_FIELD)
     }
   </td>
-  <td> <a href="@nwbib.routes.Application.show(id,"")">
+  <td> <a href="@rpb.routes.Application.show(id,"")">
     @Cache.set(session("uuid")+"-"+id, (if(i>0 && hits.size>i-1) hits(i-1) else "") + "," + (if(i+1<hits.size) hits(i+1) else ""), Application.ONE_DAY)
     @optional(doc, "title", "", "<Kein Titel>")
     @optional(doc, "otherTitleInformation", " | ", "")
