+++ app/views/tags/schema_org.scala.html
@@ -3,7 +3,7 @@
 @(doc: play.api.libs.json.JsValue)
 
 @import play.api.libs.json._
-@import controllers.nwbib.Lobid
+@import controllers.rpb.Lobid
 
 @schemaType(lobidType:String) = @{
 	lobidType match {
