+++ app/views/tags/search_advanced.scala.html
@@ -3,7 +3,7 @@
 @(label:String, q:String = "", person: String = "", name: String = "", subject: String = "", rpbspatial: String = "", rpbsubject: String = "", id: String = "", publisher: String = "", issued: String = "", sortParam: String = "", word: String = "", corporation: String = "")
 
 @import scala.collection.immutable.TreeMap
-@import controllers.nwbib.Lobid
+@import controllers.rpb.Lobid
 
 <script>
 function addSearchWidget(c){
@@ -80,7 +80,7 @@ function suggests(term){
 }
 </script>
 
-@helper.form(action = controllers.nwbib.routes.Application.search(), 
+@helper.form(action = controllers.rpb.routes.Application.search(), 
     'id -> "nwbib-form", 'role -> "form", 'class -> "form-inline") {
   @defining(TreeMap("name"->name,"person"->person,"corporation"->corporation,"subject"->Lobid.withoutUris(subject), "id"->id,"publisher"->publisher,"word"->word).filter({case (k,v) => !v.isEmpty && (k != "subject" || (k == "subject" && !v.startsWith("http")))})) { fields =>
   @for(((k,v),i)<- (if(fields.isEmpty) TreeMap("q" -> "") else fields).zipWithIndex){
