+++ app/views/tags/search_form.scala.html
@@ -2,7 +2,7 @@
 
 @(q: String, location:String = "")
 
-@helper.form(action = controllers.nwbib.routes.Application.search(q, location=location), 'id -> "nwbib-form") {
+@helper.form(action = controllers.rpb.routes.Application.search(q, location=location), 'id -> "nwbib-form") {
     <div class="input-group">
       <input type="hidden" name="location" value="@location"/>
       <input type="text" name="q" id="nwbibQuery" value="@q" class="form-control" autocomplete="off"
