+++ conf/application.conf
@@ -6,7 +6,7 @@ play.ws.timeout.idle=1200000
 play.ws.timeout.request=1200000
 
 include "../../../conf/application.conf"
-play.http.router=nwbib.Routes
+play.http.router=rpb.Routes
 
 # Secret key
 # ~~~~~
