--- conf/nwbib.routes
@@ -1,40 +1,0 @@
-# Routes
-# This file defines all application routes (Higher priority routes first)
-# ~~~~
-
-# Handle trailing slashes
-GET    /*path/                controllers.nwbib.Application.redirect(path: String)
-
-# Home page
-GET    /                      controllers.nwbib.Application.index(map?="kreise")
-GET    /info                  controllers.nwbib.Application.info()
-GET    /advanced              controllers.nwbib.Application.advanced()
-GET    /topics                controllers.nwbib.Application.topics(q?="")
-GET    /search                controllers.nwbib.Application.search(q?="", person?="", name?="", subject?="", id?="", publisher?="", issued?="", medium ?= "", rpbspatial ?= "", rpbsubject ?= "", from:Int?=0, size:Int?=25, owner?="", t?="", sort ?= "newest", details:Boolean?=false, location ?= "", word?="", corporation?="", raw?="", format?="html")
-GET    /register              controllers.nwbib.Application.register(t?="")
-GET    /classification        controllers.nwbib.Application.classification(t?="")
-GET    /subjects              controllers.nwbib.Application.subjects(t?="classification")
-GET    /subjects.ttl          controllers.nwbib.Application.subjectsTtl()
-GET    /spatial               controllers.nwbib.Application.spatial(t?="classification")
-GET    /spatial.ttl           controllers.nwbib.Application.spatialTtl()
-GET    /journals              controllers.nwbib.Application.journals()
-GET    /facets                controllers.nwbib.Application.facets(q,person?="", name?="", subject?="", id?="", publisher?="", issued?="", medium ?= "", rpbspatial ?= "", rpbsubject ?= "", from:Int,size:Int,owner,t,field,sort, location?="", word?="", corporation?="", raw?="")
-
-GET    /stars                 controllers.nwbib.Application.showStars(format?="", ids?="")
-GET    /stars/clear           controllers.nwbib.Application.clearStars(ids ?= "")
-GET    /stars/all             controllers.nwbib.Application.starAll(ids)
-POST   /stars/:id             controllers.nwbib.Application.star(id)
-DELETE /stars/:id             controllers.nwbib.Application.unstar(id)
-
-# Map static resources from the /public folder to the /assets URL path
-GET    /assets/*file          controllers.Assets.at(path="/public", file)
-GET    /robots.txt            controllers.Assets.at(path="/public",file="robots.txt")
-
-GET    /cgi-bin/wwwalleg/:name.pl   controllers.nwbib.Application.showPl(name, db ?= "rnam", index: Int ?= 1, zeilen: Int ?= 1, s1)
-GET    /sw/:rpbId             controllers.nwbib.Application.showSw(rpbId)
-GET    /o:id                  controllers.nwbib.Application.searchSpatial(id, from:Int?=0, size:Int?=25, format?="html")
-GET    /:id                   controllers.nwbib.Application.show(id, format ?= "")
-PUT    /:id                   controllers.nwbib.Application.put(id, secret ?= "")
-DELETE /:id                   controllers.nwbib.Application.delete(id, secret ?= "")
-PUT    /                      controllers.nwbib.Application.putIdFromData(secret ?= "")
-DELETE /                      controllers.nwbib.Application.deleteIdFromData(secret ?= "")
