+++ test/tests/InternalIntegrationTest.java
@@ -26,10 +26,10 @@ import org.junit.Test;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
-import controllers.nwbib.Application;
-import controllers.nwbib.Classification;
-import controllers.nwbib.Lobid;
-import controllers.nwbib.Classification.Type;
+import controllers.rpb.Application;
+import controllers.rpb.Classification;
+import controllers.rpb.Lobid;
+import controllers.rpb.Classification.Type;
 import play.libs.F.Promise;
 import play.libs.Json;
 import play.mvc.Http;
